chore: split the implementation into multiple files
