+++ brack-project-manager/src/document.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Document {
+    pub name: String,
+    pub version: String,
+    pub backend: String,
+}
